# docs/openapi.yaml
openapi: 3.0.3
info:
  title: STEPQUEST Backend API
  version: 0.2.0
  description: API contract for the current STEPQUEST MVP backend and legacy account flow.
servers:
  - url: http://localhost:3000
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
  schemas:
...
    StepQuestCreateGoalRequest:
      type: object
      required: [title, burdenLevel]
      properties:
        title:
          type: string
        category:
          type: string
          enum: [auto, study, work, writing, cleaning, exercise, wake, sleep, life_admin, relationship]
        burdenLevel:
          type: integer
          minimum: 1
          maximum: 4
        energyLevel:
          type: string
          enum: [low, medium, high]
        targetAt:
          type: string
          format: date-time
        recurrenceRule:
          type: string
          enum: [daily, weekly]
        location:
          type: string
        availableMinutes:
          type: integer
          minimum: 1
          maximum: 180
        obstacle:
          type: string
          enum: [too_big, no_material, unclear, tired, wrong_place, not_now, forgot, anxious]
    StepQuestShrinkRequest:
      type: object
      properties:
        reason:
          type: string
          enum: [too_big, no_material, unclear, tired, wrong_place, not_now, forgot, anxious, costume_active]
    StepQuestRegenerateRequest:
      type: object
      properties:
        reason:
          type: string
          enum: [too_big, no_material, unclear, tired, wrong_place, not_now, forgot, anxious]
    StepQuestReminderRequest:
      type: object
      required: [minutes]
      properties:
        minutes:
          type: integer
          minimum: 1
          maximum: 1440
        stepId:
          type: integer
    StepQuestAnyResponse:
      type: object
      additionalProperties: true
...
    EquipCostumeResponse:
      type: object
      properties:
        costumeId:
          type: integer
        equipped:
          type: boolean
        activeSkillsetId:
          type: string
paths:
  /stepquest/goals:
    post:
      tags: [StepQuest]
      security:
        - bearerAuth: []
      summary: Create a STEPQUEST goal, quest chain, and first active micro step.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StepQuestCreateGoalRequest'
      responses:
        '201':
          description: Created STEPQUEST goal state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StepQuestAnyResponse'
  /stepquest/current:
    get:
      tags: [StepQuest]
      security:
        - bearerAuth: []
      summary: Get the current active STEPQUEST micro step, user state, and village state.
      responses:
        '200':
          description: Current STEPQUEST state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StepQuestAnyResponse'
  /stepquest/stats:
    get:
      tags: [StepQuest]
      security:
        - bearerAuth: []
      summary: Get STEPQUEST attempt, step, and reward counters.
      responses:
        '200':
          description: STEPQUEST counters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StepQuestAnyResponse'
  /stepquest/dungeons:
    get:
      tags: [StepQuest]
      security:
        - bearerAuth: []
      summary: List active and paused STEPQUEST dungeons with progress, last action, and next action.
      responses:
        '200':
          description: Dungeon list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StepQuestAnyResponse'
  /stepquest/goals/{id}/pause:
    post:
      tags: [StepQuest]
      security:
        - bearerAuth: []
      summary: Pause an active STEPQUEST dungeon without losing progress.
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '201':
          description: Paused dungeon state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StepQuestAnyResponse'
  /stepquest/goals/{id}/resume:
    post:
      tags: [StepQuest]
      security:
        - bearerAuth: []
      summary: Resume a paused STEPQUEST dungeon.
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '201':
          description: Resumed dungeon state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StepQuestAnyResponse'
  /stepquest/goals/{id}/archive:
    post:
      tags: [StepQuest]
      security:
        - bearerAuth: []
      summary: Archive a STEPQUEST dungeon.
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '201':
          description: Archived dungeon state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StepQuestAnyResponse'
  /stepquest/goals/{id}/regenerate:
    post:
      tags: [StepQuest]
      security:
        - bearerAuth: []
      summary: Supersede the current chain and regenerate smaller STEPQUEST entrance steps for a goal.
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StepQuestRegenerateRequest'
      responses:
        '201':
          description: Regenerated chain and first step
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StepQuestAnyResponse'
  /stepquest/reminder:
    get:
      tags: [StepQuest]
      security:
        - bearerAuth: []
      summary: Get the saved reminder for the current user.
      responses:
        '200':
          description: Saved reminder or disabled reminder state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StepQuestAnyResponse'
    post:
      tags: [StepQuest]
      security:
        - bearerAuth: []
      summary: Save a reminder tied to the current or selected STEPQUEST micro step.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StepQuestReminderRequest'
      responses:
        '201':
          description: Saved reminder
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StepQuestAnyResponse'
  /stepquest/reminder/action:
    post:
      tags: [StepQuest]
      security:
        - bearerAuth: []
      summary: Handle a reminder action for the linked STEPQUEST micro step.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action:
                  type: string
                  enum: [complete, snooze, shrink, skip]
                minutes:
                  type: integer
                  minimum: 1
                  maximum: 1440
      responses:
        '201':
          description: Reminder action result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StepQuestAnyResponse'
  /stepquest/costumes:
    get:
      tags: [StepQuest]
      security:
        - bearerAuth: []
      summary: List STEPQUEST costume unlock, progress, and equipped state.
      responses:
        '200':
          description: Costume state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StepQuestAnyResponse'
  /stepquest/costumes/{id}/equip:
    post:
      tags: [StepQuest]
      security:
        - bearerAuth: []
      summary: Equip an unlocked STEPQUEST costume.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Equipped costume state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StepQuestAnyResponse'
  /stepquest/costumes/{id}/activate:
    post:
      tags: [StepQuest]
      security:
        - bearerAuth: []
      summary: Activate an unlocked STEPQUEST costume ability and create strategy micro steps.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Costume ability result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StepQuestAnyResponse'
  /stepquest/steps/{id}/complete:
    post:
      tags: [StepQuest]
      security:
        - bearerAuth: []
      summary: Complete the active STEPQUEST micro step and grant idempotent rewards.
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - name: idempotencyKey
          in: query
          required: false
          schema:
            type: string
      responses:
        '201':
          description: Completion result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StepQuestAnyResponse'
  /stepquest/steps/{id}/undo:
    post:
      tags: [StepQuest]
      security:
        - bearerAuth: []
      summary: Undo the most recent completion and reverse the granted rewards when no later progress depends on it.
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '201':
          description: Undo result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StepQuestAnyResponse'
  /stepquest/steps/{id}/shrink:
    post:
      tags: [StepQuest]
      security:
        - bearerAuth: []
      summary: Replace the active STEPQUEST micro step with smaller steps.
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StepQuestShrinkRequest'
      responses:
        '201':
          description: Shrink result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StepQuestAnyResponse'
  /stepquest/steps/{id}/skip:
    post:
      tags: [StepQuest]
      security:
        - bearerAuth: []
      summary: Skip the active STEPQUEST micro step without shame language.
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '201':
          description: Skip result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StepQuestAnyResponse'
  /stepquest/steps/{id}/defer:
    post:
      tags: [StepQuest]
      security:
        - bearerAuth: []
      summary: Record a stop-for-today defer attempt while preserving the active step and earned rewards.
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  enum: [not_now, tired, wrong_place, forgot, anxious]
      responses:
        '201':
          description: Defer result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StepQuestAnyResponse'
  /stepquest/steps/{id}/resume:
    post:
      tags: [StepQuest]
      security:
        - bearerAuth: []
      summary: Resume a deferred STEPQUEST micro step as the current action.
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '201':
          description: Resume result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StepQuestAnyResponse'
  /stepquest/return/eligibility:
    get:
      tags: [StepQuest]
      security:
        - bearerAuth: []
      summary: Check whether the user is eligible for a return session.
      responses:
        '200':
          description: Return eligibility
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StepQuestAnyResponse'
  /stepquest/return/start:
    post:
      tags: [StepQuest]
      security:
        - bearerAuth: []
      summary: Start a return session with a short recovery step.
      responses:
        '201':
          description: Return session result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StepQuestAnyResponse'
...
  /costumes/{id}/equip:
    post:
      tags: [Costumes]
      security:
        - bearerAuth: []
      summary: Equip an owned costume.
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Equip result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EquipCostumeResponse'